move afterMethod to BaseTest
