+++ src/test/java/ecommerce/base/BaseTest.java
@@ -2,10 +2,14 @@ package ecommerce.base;
 
 import com.aventstack.extentreports.ExtentTest;
 import ecommerce.reports.ExtentManager;
+import ecommerce.reports.ExtentTestListener;
 import io.restassured.RestAssured;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
+import org.testng.ITestResult;
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeSuite;
@@ -56,4 +60,19 @@ public class BaseTest {
         LOGGER.info(MyMarkers.TEST, "Finishing tear down of TestNG suite");
     }
 
+    /**
+     * Using dependency injection to insert ITestResult as parameter into @AfterMethod
+     * See: https://testng.org/doc/documentation-main.html#native-dependency-injection
+     *
+     * Sets the status of a SauceLabs test to pass/fail
+     * See: https://wiki.saucelabs.com/pages/viewpage.action?pageId=63472006
+     */
+    @AfterMethod
+    public void tearDown(ITestResult result) {
+        if (EnvironmentProperties.getInstance().isSauceLabs()) {
+            WebDriver driver = ExtentTestListener.getDriver(result);
+            ((JavascriptExecutor) driver).executeScript("sauce:job-result="
+                    + (result.isSuccess() ? "passed" : "failed"));
+        }
+    }
 }
