+++ src/test/java/ecommerce/tests/BrandTest.java
@@ -107,19 +107,4 @@ public class BrandTest extends BaseTest {
         return result;
     }
 
-    /**
-     * Using dependency injection to insert ITestResult as parameter into @AfterMethod
-     * See: https://testng.org/doc/documentation-main.html#native-dependency-injection
-     *
-     * Sets the status of a SauceLabs test to pass/fail
-     * See: https://wiki.saucelabs.com/pages/viewpage.action?pageId=63472006
-     */
-    @AfterMethod
-    public void tearDown(ITestResult result) {
-        if (EnvironmentProperties.getInstance().isSauceLabs()) {
-            WebDriver driver = ExtentTestListener.getDriver(result);
-            ((JavascriptExecutor) driver).executeScript("sauce:job-result="
-                    + (result.isSuccess() ? "passed" : "failed"));
-        }
-    }
 }
